Would not completey trust this build, but basics seem to be working

+++ src/main/java/au/com/teamarrow/model/Measurement.java
@@ -26,7 +26,7 @@ public class Measurement implements Serializable {
 
    private static final long serialVersionUID = -3478063846718598301L;
 	
-	private Integer id;
+	private Long id;
     private DeviceType deviceType;
     private Device device;
     private String name;
@@ -38,11 +38,11 @@ public class Measurement implements Serializable {
     public Measurement() {
     }
 
-    public Measurement(Integer id) {
+    public Measurement(Long id) {
         this.id = id;
     }
 
-    public Measurement(Integer id, DeviceType deviceType, Device device, String name, Integer canId,
+    public Measurement(Long id, DeviceType deviceType, Device device, String name, Integer canId,
         Integer reportingFrequency, String type, Set<DataPoint> dataPoints) {
         this.id = id;
         this.deviceType = deviceType;
@@ -56,11 +56,11 @@ public class Measurement implements Serializable {
 
     @Id
     @Column(name = "msrmnt_id", unique = true, nullable = false)
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
